misc update

+++ AkkuMaster/src/gde/device/htronic/AkkuMasterC4.java
@@ -184,10 +184,10 @@ public class AkkuMasterC4 extends DeviceConfiguration implements IDevice {
 		values = AkkuMasterC4SerialPort.getConvertedValues(values, AkkuMasterC4SerialPort.convertConfigurationAnswer(configurationBuffer), AkkuMasterC4SerialPort.convertMeasurementValues(measurementsBuffer));
 
 
-		points[0] = Integer.valueOf((Integer) values.get(AkkuMasterC4SerialPort.PROCESS_VOLTAGE)).intValue(); //Spannung 	[mV]
-		points[1] = Integer.valueOf((Integer) values.get(AkkuMasterC4SerialPort.PROCESS_CURRENT)).intValue(); //Strom 			[mA]
+		points[0] = (Integer) values.get(AkkuMasterC4SerialPort.PROCESS_VOLTAGE); //Spannung 	[mV]
+		points[1] = (Integer) values.get(AkkuMasterC4SerialPort.PROCESS_CURRENT); //Strom 			[mA]
 		// display adaption * 1000  -  / 1000
-		points[2] = Integer.valueOf((Integer) values.get(AkkuMasterC4SerialPort.PROCESS_CAPACITY)).intValue() * 1000; //Kapazität	[mAh] 
+		points[2] = ((Integer) values.get(AkkuMasterC4SerialPort.PROCESS_CAPACITY)) * 1000; //Kapazität	[mAh] 
 		points[3] = Double.valueOf((points[0] / 1000.0) * (points[1] / 1000.0) * 1000).intValue(); 							// power U*I [W]
 		points[4] = Double.valueOf((points[0] / 1000.0) * (points[2] / 1000.0)).intValue();											// energy U*C [mWh]
 		log.log(Level.FINE, points[0] + " mV; " + points[1] + " mA; " + points[2] + " mAh; " + points[3] + " mW; " + points[4] + " mWh"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$
